services/orders/requirements.txt
fastapi~=0.115.0
typer~=0.12.3
nats-py==2.9.0
#DB
SQLAlchemy~=2.0.35
alembic~=1.13.3
#drivers
asyncpg~=0.29.0
aiosqlite~=0.20.0
#schema
pydantic~=2.7.4
pydantic-settings~=2.2.1
